+++ blueprints/flexberry-core/index.js
@@ -7,6 +7,8 @@ var fs = require("fs");
 var path = require('path');
 var lodash = require('lodash');
 var Locales_1 = require('../flexberry-core/Locales');
+var ModelBlueprint_1 = require('../flexberry-model/ModelBlueprint');
+var CommonUtils_1 = require('../flexberry-common/CommonUtils');
 var TAB = "  ";
 module.exports = {
     description: 'Generates core entities for flexberry.',
@@ -15,6 +17,21 @@ module.exports = {
     ],
     supportsAddon: function () {
         return false;
+    },
+    _files: null,
+    files: function () {
+        if (this._files) {
+            return this._files;
+        }
+        var sitemapFile = path.join(this.options.metadataDir, "application", "sitemap.json");
+        var sitemap = JSON.parse(stripBom(fs.readFileSync(sitemapFile, "utf8")));
+        if (sitemap.mobile) {
+            this._files = CommonUtils_1.default.getFilesForGeneration(this);
+        }
+        else {
+            this._files = CommonUtils_1.default.getFilesForGeneration(this, function (v) { return v === "__root__/templates/mobile/application.hbs"; });
+        }
+        return this._files;
     },
     /**
      * Blueprint Hook locals.
@@ -75,8 +92,8 @@ var CoreBlueprint = (function () {
             var listForm = JSON.parse(content);
             var listFormName = pp.name;
             routes.push("  this.route('" + listFormName + "');");
-            routes.push("  this.route('" + listForm.editForm + "', { path: '" + listForm.editForm + "/:id' });");
-            routes.push("  this.route('" + listForm.newForm + ".new', { path: '" + listForm.newForm + "/new' });");
+            routes.push("  this.route('" + listForm.editForm + "',\n  { path: '" + listForm.editForm + "/:id' });");
+            routes.push("  this.route('" + listForm.newForm + ".new',\n  { path: '" + listForm.newForm + "/new' });");
             importProperties.push("import " + listForm.name + "Form from './forms/" + listFormName + "';");
             formsImportedProperties.push("    '" + listFormName + "': " + listForm.name + "Form");
         }
@@ -97,9 +114,7 @@ var CoreBlueprint = (function () {
             var pp = path.parse(modelFileName);
             if (pp.ext != ".json")
                 continue;
-            var modelFile = path.join(modelsDir, modelFileName);
-            var content = stripBom(fs.readFileSync(modelFile, "utf8"));
-            var model = JSON.parse(content);
+            var model = ModelBlueprint_1.default.loadModel(modelsDir, modelFileName);
             var modelName = pp.name;
             var LAST_WORD_CAMELIZED_REGEX = /([\w/\s-]*)([A-Z][a-z\d]*$)/;
             var irregularLastWordOfModelName = LAST_WORD_CAMELIZED_REGEX.exec(model.name)[2].toLowerCase();