+++ blueprints/flexberry-object-init/files/__root__/objects/__name__.js
@@ -1,4 +1,6 @@
-import { ObjectMix as <%= className %>Mixin } from '../mixins/regenerated/objects/<%= name %>';
+import Ember from 'ember';
+import { ObjectMix as <%= className %>Mixin } from
+  '../mixins/regenerated/objects/<%= name %>';
 
 let EmberObject = Ember.Object.extend(<%= className %>Mixin, {